requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astroid==3.3.5
astunparse==1.6.3
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dill==0.3.9
distlib==0.3.9
Django==5.1.3
djangorestframework==3.15.2
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.0
h5py==3.12.1
huggingface-hub==0.26.3
idna==3.10
inflection==0.5.1
isort==5.13.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.4.1
mypy-extensions==1.0.0
namex==0.0.8
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
protobuf==5.29.0
Pygments==2.18.0
pylint==3.3.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
six==1.16.0
sqlparse==0.5.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
tokenizers==0.20.3
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.27.1
Werkzeug==3.1.3
wrapt==1.17.0
